refactor: update id type from Long to Integer across entities, DTOs, repository, service, and mapper

+++ src/main/java/com/foodservice/controller/CustomerController.java
@@ -22,7 +22,7 @@ public class CustomerController {
 
     // Get customer by id
     @GetMapping("/{id}")
-    public CustomerDTO getCustomerById(@PathVariable Long id) {
+    public CustomerDTO getCustomerById(@PathVariable Integer id) {
         return customerService.getCustomerById(id);
     }
 }
